+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.6.5
 `[NEW]`
 - Added interactive bash auto completions.
-- Added flag `--check-version`, this will read your changelog and check for new version.
+- Added flag `--update-ascii-version`, this will read your changelog and check for new version.
 
 `[IMPROVEMENTS]`
 - Improved help info.